+++ packages/firebase_auth/android/src/main/java/io/flutter/plugins/firebaseauth/FirebaseAuthPlugin.java
@@ -83,6 +83,9 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       case "signInWithFacebook":
         handleSignInWithFacebook(call, result);
         break;
+      case "signInWithTwitter":
+        handleSignInWithTwitter(call, result);
+        break;
       case "signOut":
         handleSignOut(call, result);
         break;
@@ -240,6 +243,15 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
     Map<String, String> arguments = (Map<String, String>) call.arguments;
     String accessToken = arguments.get("accessToken");
     AuthCredential credential = FacebookAuthProvider.getCredential(accessToken);
+    firebaseAuth
+        .signInWithCredential(credential)
+        .addOnCompleteListener(new SignInCompleteListener(result));
+  }
+
+  private void handleSignInWithTwitter(MethodCall call, final Result result) {
+    String authToken = call.argument("authToken");
+    String authTokenSecret = call.argument("authTokenSecret");
+    AuthCredential credential = TwitterAuthProvider.getCredential(authToken, authTokenSecret);
     firebaseAuth
         .signInWithCredential(credential)
         .addOnCompleteListener(new SignInCompleteListener(result));