+++ packages/firebase_auth/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.9
+
+* Added support for signing in with a Twitter account.
+
 ## 0.5.8
 
 * Added support to reload firebase user